Add files via upload
Title: Titanic Classification
Description: This file contains the code for a machine learning model that predicts the survival of passengers on the Titanic. The model is trained on a dataset of passenger information, including their age, sex, passenger class, and fare. The model is then used to predict the survival of passengers in a test dataset.
Algorithms used: The model uses the logistic regression algorithm.
Data: The data used to train and test the model is the Titanic dataset, which is available on Kaggle.
Libraries used: The following libraries are used in this file:
pandas
numpy
matplotlib
seaborn
sklearn
Instructions: To use this file, first install the necessary libraries. Then, open the file in a text editor or Jupyter notebook. You can then run the code to train and test the model.