TASK-3_handwritten-digit-recognition.ipynb
Title: Number Recognition
Description: This file contains the code for a machine learning model that recognizes handwritten digits. The model is trained on a dataset of handwritten digits, such as the MNIST dataset. The model is then used to recognize handwritten digits in a test dataset.
Algorithms used: The model uses a convolutional neural network (CNN).
Data: The data used to train and test the model is the MNIST dataset, which is available on Kaggle.
Libraries used: The following libraries are used in this file:
keras
tensorflow
numpy
matplotlib